Revert "Feature/task/ignore errors (#48)"
This reverts commit d85c2b8dc151ab4ece15cecdd2ce72c686a00604.
