+++ examples/assert.rh
@@ -10,4 +10,3 @@
   assert:
     that:
       - "rash.path == ''"
-  ignore_errors: true
