+++ rash_core/src/utils/mod.rs
@@ -1,5 +1,4 @@
 pub mod file;
-pub mod jinja2;
 
 use crate::error::{Error, ErrorKind, Result};
